CGED prototype
basic turned base game for CGED. the Action class is used to establish each attack by giving it a button and an attack type. the class used to have a status effect parameter, but this meant that the enemy was unable to access status effects (player actions are handled differently). each role is is initialised with the Role class so that they have their own separate stats and attacks they can do. as of right now, the player has 3 attacks (light, medium, heavy) and the heavy attack can only be used every 5 turns. thinking of reducing number of attacks to 2 or giving one of the attacks a special purpose that is better suited for certain situations (not sure what this could look like). the player is able to use magic that can apply different status effects (these have not been implemented yet). the player is also able to use potions to restore health. the enemy currently selects attacks randomly from a preset list based on what role it is. role is also randomly selected each time the game is loaded. enemy is unable to apply status effects as their attacks are not Actions. currently changing code to hopefully make it work. magic attacks will be turned into dictionaries (can't remember how i will continue this approach atm)