updated how buttons and attacks are stored
Actor and StatusBar classes have been moved out of the main file. Action class is no longer needed as I came to the conclusion that dictionaries would be more appropriate. All attacks have been turned into dictionaries and I am now in the process of changing the main code to make use of them. Also  moving code around and deleting variables that don't need to be there. Player attacks are functional, but they have access to magic attacks not found in their role, will be changing the Role and player attack loop to make use of new dictionaries. Will also change the arguments of the Role class to make it clear that attacks are for both player and enemies